Logout button updated
